chore(client): remove unused files and restructure components for better organization and maintainability
