+++ client/src/main.tsx
@@ -1,4 +1,4 @@
-import {Page} from '@/app/dashboard/page.tsx';
+import {Page} from '@/pages/page.tsx';
 import {StrictMode} from 'react';
 import {createRoot} from 'react-dom/client';
 import './index.css';
